+++ app/models/user.rb
@@ -28,6 +28,7 @@ class User < ActiveRecord::Base
   has_many :dataset_file_audits
   has_many :tools
   has_many :topics, -> { where deleted: false }
+  has_many :subscriptions
 
   # User Methods
 
@@ -48,11 +49,14 @@ class User < ActiveRecord::Base
   end
 
   def subscribed_topics
-    Topic.current
+    Topic.current.not_banned.where( id: self.subscriptions.where( subscribed: true ).pluck( :topic_id ) )
   end
 
+  # All comments created in the last day, or over the weekend if it is Monday
+  # Ex: On Monday, returns tasks created since Friday morning (Time.now - 3.day)
+  # Ex: On Tuesday, returns tasks created since Monday morning (Time.now - 1.day)
   def digest_comments
-    Comment.digest_visible.where( topic_id: self.subscribed_topics.pluck(:id) )
+    Comment.digest_visible.where( topic_id: self.subscribed_topics.pluck(:id) ).where("created_at > ?", (Time.now.monday? ? Time.now.midnight - 3.day : Time.now.midnight - 1.day))
   end
 
   def all_datasets